Define anomaVerify in terms of anomaVerifyWithMessage (#3)
This change is made to avoid adding an additional builtin function to
the Juvix compiler as it can be simply implemented in terms of
`anomaVerifyWithMessage`.

The polymorphism of `anomaVerifyWithMessage` requires us to apply a
type. Any type will do here because the value is not used. I've chosen
`Unit` because then it's clear to readers of the source code that the
result of `anomaVerifyWithMessage` is being discarded.

This commit also adds `isJust` which should be added to the Juvix
standard library along with other utilities for the Maybe type.