Improve type safety of common Anoma types (#11)
* Improve type safety of common Anoma types

Previously we used Nat for almost all types (PublicKey, Commitment,
Nullifier etc.) because the Anoma builtin functions and the Anoma Node
interface requries it.

Howver, the Juvix Anoma backend represents Juvix records as Nockma
tuples. In particular records with a single field (i.e newtypes) are
represented by their wrapped value. Therefore using a newtype instead of
the wrapping type does not change the runtime representation in Nockma
so we can freely use them in this library.

If we used the newtypes in the Anoma builtin function signatures then we
would have to make the newtypes builtin too in order that they could be
supported in the JuvixCore evaluator. This is undesirable because we
would have to update the compiler's Anoma backend whenever we changed
these newtypes.

Instead we define wrapper functions in `Anoma.System` which delegate to
the builtins in `Anoma.System.Builtin`. The functions in
`Anoma.System.Builtin` are not exposed in the main `Anoma` module. The
wrapper functions are inlined to avoid adding a function call at
runtime.

* Remove unused function

* Reformat project